Revisit nonlinear subdivision in vtkDataSetSurfaceFilter
Some weeks ago, the cell->triangulate function was updated to
return local point ids instead of global ids. Today, we take
advantage of this recent change to use local ids in the nonlinear
subdivision algorithms as well.
This has several advantages:
* Before, we had to track both the output ids and the parametric
points ids. This involves a lot of data copies. Now, the algorithm
is only applied to local ids.
* This also allows us to remove some duplicated searches in the
map PointMap.
* The single desavantage is that  when the subdivision level is
higher than 1, a new local vtkEdgeInterpolationMap is used. But
this map is small, and avoids numerous calls to the global maps
EdgeMap and PointMap.
I have monitored the computation cost of that function.
The speed is improved if the subdivision level is 1.
The speed is the same if the subdivision level is higher than 1,
because in that case the computation cost is driven by the number
of evaluations, which is unchanged.

But the main motivation of this change is that working with local
ids will help in supporting nonlinear subdivisions of degenerate
cells.