Add the Alembic ".abc" format scene exporter.
Alembic is a scene description, and exports the current scene's
rendered appearance.

Polydata include texture coords and vertex
colors, if available, and LUTs are exported as a texture .png file
with the same file root, and "_texN.png" suffixes.

The active camera includes its transform and clipping planes.

Lights are not currently supported by Alembic, so are not included.