Catalyst add channel field data at partition level
* Conduit node `catalyst/channel/<channelName>/data/state/fields/`
contains field data. Add the field data at the leaf level of the
PartionedDataSet, i.e. on the concrete dataset.
This used to be added only on the root level, and thus appears
as `partial` in paraview with only one rank used for the value.

Now each partition has its own value, and array is not partial anymore