Fix parsing of literal operator id suffix
Previously, the wrapper parser failed with a syntax error if
operator""_xyz did not have a space between the "" and _xyz.
The space was required in C++11, but was made optional in later
versions of C++.

Also remove the requirement that the user-defined suffix begins
with an underscore, since this isn't required by C++ syntax,
which simply defines "ud-suffix" as "identifier".