vtkTableBasedClipDataSet: Optimize clip tables and remove most switch-cases
Tables were handling both cases when insideout is on and off using color.
Also, there was a separate table for determining the number of outputs shapes
of each case. and non ST_PNT outputs, a.k.a. normal cells did not have the
number of points present so a switch case was needed. Finally, centroidIdx
was present but always 0.

Now there 2 tables, one for insideOut == false and one for insideOut == true.
Each one is accessed based on the template insideOut parameter of the tables.
Either of these tables include for each case,

the number of output cells
1. shape, number output points, point ids p0, p1, ...
2. ....

There is a helper function to get the case given a vtk cell type and caseIndex.
Finally, there is a helper function to get the vtk cell type from shape.