Fix for reading binary XML files > 2Gb on Windows
Fix an error encountered when reading binary format XML files
on Windows - a missing define for Expat caused the offset used reading
the position in the file to be limited to 2Gb. Fix the define in internal Expat,
and add a check in `CanReadFile()` when using external Expat that the
necessary feature, `XML_LARGE_SIZE`, is enabled. Add a call to `CanReadFile()`
in the large-file test. Re-enable the text on Windows.