Fix line widget translation interaction bug
When moving a line widget with one of its axes at 0, it can lead to
matrix calculation inaccuracies. Theses inaccuracies result in a computed
value close to zero, but not precisely zero as it should be.
This slight offset prevent some filters relying on an intersection between
a line and a plane to works correctly.

Fix this bug by projecting the new point onto a plane that is defined
by the camera's orientation axis and a reference point within that plane.

(cherry picked from commit ccda4c2453c56a40ace76d27f759dcdb2e34e48e)