vtkRedistributeDataSetFilter: fixing handling of ghost points and cells
`vtkRedistributeDataSetFilter` was removing ghost cell and ghost point
arrays. This was not the correct behavior, as the filter can produce
duplicate cells. What needs to be done is to remove the duplicateness of
ghost points in the output, as the interface between partitions changes,
making them obsolete and wrong.

Adresses #22340