Fix Java error with integer, enum overloads
The Java wrappers simply cast enum types to 'jint', and do not
distinguish them as individual types that are distinct from 'int'
or 'unsigned int'.  So when a method is overloaded with both 'int'
and an enum type, only one of the overloads can be wrapped.  The
'int' (or 'unsigned int') is the one that is chosen.