Fix TestPHyperTreeGridProbeFilter
The vtkPHyperTreeGridProbeFilter parallel filter is now tested
using a vtkRandomHyperTreeGridSource as input, which can be distributed.
Changes around the rendering pipeline to support the distributed
pipeline are also made. Finally, the sequential versio of the test
is also modified to use the vtkRandomHyperTreeGridSource as well and
ensure that both tests produce the same output.

(cherry picked from commit 45c21489316bae5534c792fce54835b4528037a7)