vtkADIOS2CoreImageReader: respect Image Dimension array
The Image Dimension property is used to determine the dimensions of the
full dataset, however, when determining the dimensions of individual
blocks, the first array in the list of enabled arrays were being used.
This causes a bug in situations where different arrays have different
block dimensions, even though the dimensions of the full data are the
same across arrays.