vtkADIOS2CoreImageReader: better recovery from error state
Changing Image Dimension array to an invalid array caused a rest of the
ADIOS objects, however, the state tracking if metadata has been read was
not reset leading to a crash when you later choose an invalid array.
Selecting an incorrect array name doesn't affect the ADIOS objects,
resetting is unecessary. Instead we'll just provide an empty multiblock
and leave the ADIOS objects alone.