Fix wrong sign in vtkImplicitPolyDataDistance (Issue #18307)
Fix a bug in vtkImplicitPolyDataDistance that returns the correct distance with
the wrong sign under certain conditions, such as when the nearest point is on
the edge of a cube. This fixes the first problem reported in issue #18307 by
cleaning the input polydata, so that nonunique points are merged and different
faces of the cube share the same points.

(cherry picked from commit e15b0af222dcd4ba0c27fbb1726cad33e114c40f)