BUG: Recognize OpenVR gesture where buttons are pressed consecutively
This commit introduces an additional action named `ComplexGestureAction_Event2`
to address the recognition of complex gestures in OpenVR. In cases where two
complex gesture buttons are pressed consecutively, the `actionData.bChanged`
condition (utilized in `DoOneEvent`) is already false when the user is already
holding down one button, leading to the absence of additional event synthesis.

By introducing `ComplexGestureAction_Event2`, a second event is synthesized,
resulting in a second invocation of `HandleComplexGestureEvents`. This enables
the function `RecognizeComplexGesture()` to effectively identify and recognize
the associated gesture.

Co-authored-by: Lucas Gandel <lucas.gandel@kitware.com>