VR: Provide stereo projection transforms for rendering
This commit addresses proper rendering, particularly by
`vtkOpenGLGPUVolumeRaycastMapper`, when the head-mounted display (HMD)
is positioned inside the volume. The issue is resolved by ensuring
correct computation of the view frustum when the camera is tracking the
HMD and is positioned inside the volume.

The computation of the view frustum is adapted to account for the left
and right eye matrices explicitly created for VR. Note that the updated
logic adjusts the frustum computation: it utilizes the left/right
matrices when tracking the HMD, and defaults to the regular frustum
computation method based on the viewport aspect ratio if hardware
picking is enabled. In hardware picking scenarios, the camera's
viewpoint is aligned with the wand controller, and TrackHMD is disabled
to maintain stability and prevent unintended scene changes.

Co-authored-by: Lucas Gandel <lucas.gandel@kitware.com>
Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>