VR: Move the TrackHMD flag to vtkVRCamera
Track HMD affects whether the camera is tracking the head mounted
display or not. Moving this flag to the vtkVRCamera is fitting for this
purpose to allow multiple cameras in the VR world.  When performing
hardware picking, the camera's viewpoint is aligned with the wand
controller and TrackHMD is disabled to maintain stability and ensure
that the scene doesn't change.