Add CI tests for ABI Mangling feature
Add a build/test pair that only tests the ABI mangling.

The tests for this feature only run the ABI mangling tests and
excludes the rest of the test suite. The rest of the test suite
may be enabled at a later date when failing tests are addressed
in conjunction with the ABI mangling feature.