VR: Declare AddAction() functions as virtual
This commit virtualizes the `AddAction()` functions in the VR
interactor. The addition of the `isAnalog` parameter aims to facilitate
client applications that support both OpenXR and OpenVR runtimes. Although
the `isAnalog` parameter is ignored in the context of the OpenXR interactor,
these signatures are designed to meet the requirements of the base class
interface.