Do not always clear the renderer in vtkSSAOPass
The Transparent flag for colors, and the PreserveDepthBuffer flag for depth
must be checked before clearing the renderer.
In addition, clearing is already handled in the camera pass of the delegate
pass, so this code should only be executed for the built-in
vtkRenderer::UseSSAO approach where the delegate is a simple opaque pass.