vtkCameraOrientationWidget: Avoid calling Render when state is the inactive
When moving the mouse, the interaction state of the representation was
computed, then the widget state was set, and later on the representation
was refreshed and a Render was called. This is not ideal because if the
widget state is inactive, a render would be called regardless. Now a
render is called only when the widget is not inactive.