Rendering: Fix stereo image separation bug
Fix how the final transformation, T, described in "Generalized Perspective
Projection" is applied when doing off-axis projection.  This transform
encapsulates the viewpoint offset, and should be computed using the same
eye positions used to compute the rest of the projection matrix, but was
inadvertently computed using the single "EyePosition".

The result of this bug was that stereo images of an object were always a
fixed distance apart from each other, regardless of the objects location
w.r.t. the screen, and regardless of the the eye position w.r.t the object
and screen.

This commit moves computation of the left/right eye position into a new
method "ComputeEyePosition" and uses that method from both the projection
and view transform computations.  It also adds test of stereo image
left/right offset.