Particle tracer: make it inherit vtkTemporalAlgorithm
`vtkTemporalAlgorithm` embeds time series managenement that is
compatible with stock VTK or in situ using Catalyst. The particle
tracers subclassing `vtkParticleTracerBase` managed time in
a very convoluted way, causing a memory leak when running in situ.
Time management is now entirely delegated to its superclass, which
makes the code more maintainable and easier to understand.

The P filters have been deprecated. MPI logic is now embedded
in the base classes.

Some API has been deprecated, specifically the API letting someone to
change the start time or termination time of the simulation. It was not
tracktable to add this logic layer in `vtkTemporalAlgorithm`, and one
can already subsample the time series by modifying the information key
`TIME_STEPS` in `RequestInformation`.

`vtkParticlePathFilter` seemed not to work at all. ParaView has its own
custom version of this filter (`vtkInSituParticlePath`), which can now
be replaced by the VTK`s one, as it now works as expected.