rendering: Fix gpu ray cast volume mapper
Fix a problem with the mapper that breaks volume rendering when
the active vtkCamera has a non-identity ModelTransformMatrix.

Since the eye position is needed in data space, and the camera's
ModelViewTransform can affect that, combine camera and volume
matrices to infer the eye position for each volume.  Also, this
can be computed on the cpu, avoiding a matrix/matrix multiply and
and handful of matrix/vector multiplies in the fragment shader.

(cherry picked from commit cd252fcbf6b651584102b70391bf5246ca62e38b)