Improve accuracy of vtkWindowedSincPolyDataFilter
Using Hamming windowing function resulted in inaccurate scaling of the output
(up to about 1% error, either increase or decrease of the scale).

This commit implements all three window functions that are described in Taubin's paper and adds Nuttal window
and changes the default to Nuttal to eliminate the scaling error.

Note that when passband is set to very small value then shrinkage may occur regardless of the choice of window function.