fix(ResampleWithHTG): fix wrong resampling with HTG and MPI.
The `vtkPResampleWithDataSet` could produce incorrect values when having a `HyperTreeGrid` as a source and using multiple process. It was due to the fact that the `vtkCompositeDataProbeFilter` makes an internal call to `vtkHyperTreeGridProbeFilter`  but didn't propagate the value of the MaskPoints, which led to no point data being exchanged between any of the process.

(cherry picked from commit b06be9998773527049e152cec99277f8b203e553)