feat(DecimatePolyline): add a new strategy based on custom field.
Add a new decimation strategy based on a custom field array passed by the input, this array is a PointData and contains values that will be used to perform the decimation. Currently it will compute the eulerian distance between each component of the array, for each 3 vertex considered by the decimation.