vtkReflectionFilter: fix cell point ids in reflected triangle strips
With CopyInput on, triangle strips with an even number of triangles
were not reflected correctly. The first four cell ids were not set
correctly because the number of points in the copied input was not
added to these point ids, so they pointed to points in the copied
input, not the reflected points. Fix that by computing a point id
offset that depends on the CopyInput property and applying it to all
cell ids when reflecting triangle strips that match this criteria.