vtkIOSSReader: Release Regions only when needed
Previously regions were released at the end of every
ReadMesh to solve issue paraview/paraview/-/issues/21340.
This caused the creation of regions at each ReadMesh
which requires re-reading their metadata.

In this commit, we release the regions only if we are
about to read new restart file(s). This way we solve the
memory problem of paraview/paraview/-/issues/21340 and
then performance problem of re-reading metadata multiple
times of paraview/paraview/-/issues/22572.