Turn warning into TRACE log to limit its visibility
Introduced and discussed in MR !11026

This was set as a warning because data allocation
is against the implicit array philosophy. But there are
too many legetimate code paths that goes through
it (like rendering),
so turn it into a Log with higher verbosity.

Also see #19304 for an example of code path in
python wrapping.