Add guard to zSpace-related render windows
zSpace render windows are dedicated to stereo rendering on zSpace Inspire hardware.
We expect to use them with StereoRender enabled. If not, they will display nothing.
But we need to add a check to prevent the SubmitFrame method to be called in that case.
This method (zSpace API) can crash on some circonstances (for example when going fullscreen)
when StereoRender is not enabled. In the future, further investigations could be done to
handle such case in a better way.