Add vtkRemoveGhostCells filter
Add the vtkRemoveGhostCells filter to remove ghost cells and ghost array of
the input HTG.

Also update the `ghost.htg` testing data, that has been generated with previous version
of HTG ghost cell generator. It appears that this version was generating cells with undefined
ghost / mask information (visible when removing ghost cells).
The new data has been generated with the new version and do not have this issue.