Added 'vtkOFFReader' for Geomview OFF file format
This commit implements a new reader class which can import the
commonly used Geomview Object File Format (or 'OFF' in short),
resulting in a vtkPolyData object. The skeleton code of the reader
class is based on the vtkOBJReader class.

This reader only supports the simple ASCII version of the OFF
format without extensions to higher dimensions or colors, however,
the reader is capable of ignoring comment lines.

A corresponding self-contained unit-test has also been added.

More details on the OFF format can be found here:
* http://www.geomview.org/docs/html/OFF.html
* https://en.wikipedia.org/wiki/OFF_(file_format)