fix(RandomHTGSource): make the distribution more appealing with large masks.
This brings a distribution that favorizes the unmasking of the small cells. The old one led to having only very large cells unmasked with a MaskedFraction > 0.5. To do this we added multiple shuffles to remove the biais on our algorithm, which tends to mask the first trees it processes. We also used the level of the node to ponderate weither or not it should be masked, the deeper the node is (the smaller it is), the less the chances are of it to being masked. Also we had to change the seed depending on the piece to avoid having observable structures repeating in distributed mode.