Multitouch interaction and gestures
Changes the logic of gesture processing slightly. Originally, once a
gesture was detected, VTK would continue processing it for all future
mouse move (touch move) events until the finger was lifted or mouse was
released. This is fine for single touch gestures but for multitouch
gestures, it is typically required to switch to different gestures with
different motions. For example, the user could touch the screen, start
pinching and without lifting the finger, start panning. Another common
interaction is when the user starts panning but the pointer math
initially guesses the gesture to be a pinch gesture. Originally
execution would be stuck processing pinch even if the user's fingers are
performing a pan. This change would ensure that gesture recognition
eventually matches up with the user's intention.