Fix crash in vtkPointLocator when number of points is changed to zero
When the number of points in the input dataset was set to zero, BuildLocator would be called but the HashTable would not be invalidated.
The FindClosestPoint methods would continue on with the invalid HashTable, resulting in a crash when invalid point IDs were accessed.

Fixed by calling FreeSearchStructure before returning from BuildLocatorInternal so that the HashTable is invalidated.
FreeSearchStructure now also re-initializes variables that are calculated during BuildLocatorInternal so that Bounds/H/Divisions will be consistent for an empty dataset.
Removed vtkErrorMacro when the dataset is empty or has no points since it is a valid input.

(cherry picked from commit c14d87f862ed16ebabed3bfb61464216db269602)