ImplicitTypedArray::NewInstance should return AOSDataArray
ImplicitTypedArray's are the vtkImplicitArray (like vtkConstantArray)
specialized class generated at configure time for standard value types
(e.g. vtkConstantIntArray).

As `NewInstance` is expected to produce a writable array,
those classes should return a vtkAOSDataArrayTemplate of the
same value type instead of a vtkImplicitArray.

For this exact purpose, there is a `vtkImplicitArrayTypeMacro`
that should be used instead of the usual `vtkTypeMacro`.