BUG: DICOMParser: UT uses 4-byte value lengths, not 2-byte ones
Section 7.1.2 in Part 5 of the DICOM Standard [1] defines a list of VRs
that use 2-byte lengths instead of the default 4-byte ones, but UT is
not part of that list. Using 2-byte lengths for UT therefore caused the
parser to break with any DICOM file that contains UT records in its
header.
See pydicom [2] for an alternate implementation that also uses 4-byte
lengths for UT.

[1]: https://www.dicomstandard.org/standards/view/data-structures-and-encoding#sect_7.1.2
[2]: https://github.com/pydicom/pydicom/blob/abbacee6f20428930037cadf49aef94969617767/src/pydicom/valuerep.py#L486-L591