Remove loading extension GL_ARB_gpu_shader5 in Volume vertex shader
The extension was added in 3dd6f6bc6 for OpenGL 1.2 compatibility with the inverse function.

This function was used in multiple places then, but now only used in the Fragment shader clipping (VTL::Clipping::Init), making the inclusion of the extension irrelevant.
Loading this extension did cause shader compilation issues in GLES, used especially in the WASM build. If this change causes issues for older OpenGL, enable the extension conditionnally when GL_ES is not defined