vtkHyperTreeGrid: fix casts in asserts
Seems like the casts were improperly fixed in bf068ae0fd (Fixed various
debug-only -Wsign-compare warnings, 2024-04-23) (merged via !11085). The
`local_ijk` is an array of `unsigned`, so the cast of `-dN` needs to be
unsigned as well. This is in-range because `dN` is a negative number at
this point of the `assert`.