GLES Volume rendering: use NEAREST instead of LINEAR mag texture filtering
The OpenGL ES 3.0 specification defines a texture as incomplete when 'a sized internal color format that is not texture-filterable, and either magnification filter is not NEAREST [...]'. This is is what the Volume Rendering DepthTextureObject used, so the magnification had to be changed to NEAREST for it to work in OpenGL ES (through WebAssembly/emscripten for example).

(cherry picked from commit 216aca089319cd571255471cd57c7f6fe78469d6)