Add static ComputePhysicalToIndexMatrix method
There are a couple advantages to having a dedicated function for
computing the PhysicalToIndexMatrix rather than simply inverting
the IndexToPhysicalMatrix.

1. It's cheaper, since it only needs vtkMatrix3x3::Invert() and
not the much more expensive vtkMatrix4x4::Invert().

2. If the IndexToPhysicalMatrix is a singular matrix due to one or
more spacing values equal to zero, we zero the corresponding row
(or rows) in the PhysicalToIndexMatrix, so that IndexToPhysical
and PhysicalToIndex matrices are pseudoinverses of each other.