Fixes and comments for vtkImageTransform
More operations are done up-front to make the inner loops tighter.
Checks have been added to avoid division by zero.  The inverse
transpose matrix is now used consistently in all normal and vector
transformation code paths.

Warning statements have been added that vectors are assumed to
be gradients and are therefore transformed as covariant vectors.
This is a valid assumption for the filters this class is used with,
but it's important to warn people that this isn't a general-purpose
transform class.