Fix occlusion culling incorrect behavior
Extended the number of mipmaps up to 1x1 which may be needed for very large objects (the shader was reading in undefined mipmap level before this fix, leading to undefined behavior and incorrect culling on some drivers).

Proper redefinition of the number of mipmaps of the hierarchical z buffer texture view in the culling pass of the occlusion culler (prior to this commit, the texture view number of mipmaps wasn't adjusted for a larger/smaller number of mipmaps after a window resize).