Force order of unregistration of cell-grid responders.
Cell-grid responder classes often hold other VTK objects.
If the registrar holding responders happens to be destroyed
after the vtkDebugLeaks or the vtkFilteringInformationKeyManager
classes perform their finalization, this can result in either
false-positive leaks or crashes (when information-keys that
have been deleted are dereferenced) or both.

Add vtkCellMetadata::ClearResponders() to force all cell-grid
responders to be unregistered and ensure it is called before
vtkDebugLeaks::ClassFinalize or
vtkFilteringInformationKeyManager::ClassFinalize are run.