Use WritePointer in vtkMultiBlockPLOT3DReader
When getting a naked pointer for writing to an array, the
WritePointer() method ensures the array is adjusted so that it
can receive the new values.  If just GetPointer() is used, then
the MaxId of the array is not correctly set.

(cherry picked from commit 2ac14b21906525594c02a59fc5442f92059a01cd)