Fix array write beyond MaxId in vtkSMPMergePoints
Calling SetNumberOfPoints()/SetNumberOfTuples() on the points and
arrays before merging the results ensures that we never index an
array past the MaxId of the array.  We call Resize() first, since
Resize() will preserve existing data on reallocation, while
SetNumberOfTuples() won't preserve data on reallocation, but
will preserve data if the array stays the same size or shrinks.