vtkXOpenGLRenderWindow: Support borderless and coverable windows
There does not seem to be a modern way to make windows borderless
but also not force them to be always-on-top.  This change uses old
motif window manager hints to achieve this goal.  While motif is
ancient, it seems most window managers still support these hints,
and will continue to do so as long as X11 is around.