Add 'encoding' to vtkPythonInterpreter IO streams
When run in interactive mode under Python 3.13, the vtkpython exe
raised "AttributeError: 'vtkPythonStdStreamCaptureHelper' object has
no attribute 'encoding'" whenever it tried to print to the screen.

Adding an attribute "encoding='utf-8'" to the stream capture objects
fixes the exception.  Note that there is no check that the console is
actually UTF8, but VTK has historically assumed that it is.