vtkOpenGLBatchedPolyDataMapper: fix inconsistent rendering for composite
Previously, the DrawIBO depends on the order of the VTKPolyDataToGLBatchElement
map. As this map is ordered by the address of the dataset which change in
a context of composite temporal data, it results of inconsistent
rendering when playing an animation.

Now, the order depends on the flatIndex produced by the caller.