vtkOpenGLBatchedPolyDataMapper: fix inconsistent rendering for composite
Previously, the DrawIBO depends on the order of the VTKPolyDataToGLBatchElement
map. As this map is ordered by the address of the dataset which change in
a context of composite temporal data, it results of inconsistent
rendering when playing an animation.

Now, the order depends on the flatIndex produced by the caller.

(cherry picked from commit 5b04869c2d6506d616481380f7c2a3c4654fe6f3)