Fix int overflow and perf improvements
The tetrahedron reading logic used a non-tetrahedron-specifc `vtkUnstructredGrid::InsertNextCell` call which performed poorly due to a large amount of time spent managing a `std::set`, and struggling with memory allocation previsions.

With this change we instead handle the cell ID uniqueness ourselves and use the tetrahedron-specific API, resulting in a 2~3x speedup on a large dataset.

The logic is also refactored a bit for readability

Fix an int overflow that could occur when seeking a very large offset (it did happen in the geometry file)

Overall performance improvements