vtkContour3DLinearGrid: quiet an INFO message
An INFO message is always printed with default log verbosity. This
INFO message was in a function that checks a capability, so could be
logged during the course of an entirely normal operation. This changes
it to TRACE to quiet it but still make it available at higher
verbosity levels.

(cherry picked from commit 894fa2de1b4e1d404fd34c22346fedf8806d0735)