Continue to process metadata even if paths is empty
With [recent changes to Fides], it is no longer necessary to supply
paths when loading a JSON file (as long as the JSON gives a file path
relative to the JSON path). Because this case can happen, the
`RequestInformation` of the Fides reader should continue to process the
meta data even if the paths object is empty.

This is particularly important in ParaView to populate the field
selection before they `Apply` button is hit. Otherwise, all the fields
must be loaded before they can be deselected.

[recent changes to Fides]: https://gitlab.kitware.com/vtk/fides/-/merge_requests/172