Change X11 event loop termination for multi-window
The Win32 and Cocoa interactors had an event loop that would
break if TerminateApp() or SetDone(true) was called on any
window's interactor, but for the X11 event loop, it was necessary
to call TerminateApp() or SetDone(true) on the interactor of
all the windows.  This change makes X11 behave the same as Cocoa
and Win32 (which is also the behavior it had in VTK 8, when the
X11 interactors used an XtAppContext singleton).