vtkLog: sweep to change INFO log messages to TRACE
INFO log messages will be logged by default to stdout, but this
logging level was used in a number of places throughout VTK, leading
to unexpected output messages. VTK should not print messages to stdout
by default except when reporting a warning or error condition. This
commit changes messages that do not report warnings/errors to TRACE so
messages will be logged only when the logging level is specifically
requested to be more verbose.