Fix Cache Status when consumer changes data type
One cannot compare Cache and Original data type:
the consumer can change it! E.g. ExtractSurface
can have an UG as Original data but has a PolyData as Cache.

What we want to check here is only input type
across different executions. So put the check
in SetOriginalData instead and invalidate the
Cache as needed